requirements.txt
opencv-python
pyautogui
tensorflow-gpu==2.10
numpy==1.19
PyQt5
SpeechRecognition
pyttsx3